Moved the Create Button Modal into ManageAccountPage
