+++ packages/frontend/src/pages/routes/AuthenticatedRoutes.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import { Routes } from 'react-router';
 import { Navigate, Route } from 'react-router-dom';
 import HomePage from '../public/HomePage';
-import CreateHousePage from '../public/CreateHousePage';
-import ManageAccountPage from '../public/ManageAccountPage';
+import ManageAccountPage from '../private/ManageAccountPage';
 
 interface AuthenticatedRoutesProps {}
 
@@ -13,7 +12,6 @@ const AuthenticatedRoutes: React.FC<AuthenticatedRoutesProps> = () => {
       <Route path="home" element={<HomePage />} />
       <Route path="*" element={<Navigate to="/home" replace />} />
       <Route path="account" element={<ManageAccountPage />} />
-      <Route path="create" element={<CreateHousePage />} />
     </Routes>
   );
 };
